Moved BodyType out of BodyReader.
