+++ rawhttp-core/src/main/java/rawhttp/core/body/BytesBody.java
@@ -3,7 +3,7 @@ package rawhttp.core.body;
 import java.io.ByteArrayInputStream;
 import java.util.OptionalLong;
 import javax.annotation.Nullable;
-import rawhttp.core.BodyReader;
+import rawhttp.core.BodyType;
 import rawhttp.core.LazyBodyReader;
 
 /**
@@ -30,7 +30,7 @@ public class BytesBody extends HttpMessageBody {
 
     @Override
     public LazyBodyReader toBodyReader() {
-        return new LazyBodyReader(new BodyReader.BodyType.ContentLength((long) bytes.length),
+        return new LazyBodyReader(new BodyType.ContentLength((long) bytes.length),
                 null,
                 new ByteArrayInputStream(bytes));
     }
